FUNFUNZ-118: moved all hooks from service to hooks folder
