+++ src/services/entity.ts
@@ -1,8 +1,6 @@
-import { useEffect } from "react"
-import { dispatch, useSelector } from "reducers"
+import { dispatch } from "reducers"
 import { FETCH_ENTITY_FULFILLED, FETCH_ENTITY_PENDING, FETCH_ENTITY_REJECTED } from "reducers/entity"
 import graphql from 'services/graphql'
-import { IField, mapFieldComponents } from "utils/fields"
 
 export interface IProperty {
   name: string,
@@ -79,33 +77,4 @@ export async function getEntity(entityName: string): Promise<IEntity> {
       throw error
     }
   )
-}
-
-/*
- * To be used on Create, View and Edit page to render input fields
- */
-export function useEntity(entityName: string): IEntity & { label: string, fields: IField[] } {
-  const entity = useSelector((state) => state.tables.find(t => t.name === entityName))
-  const loading = useSelector((state) => state.loadingTables || state.tables.find(t => t.name === entityName)?.loading)
-  const error = useSelector((state) => state.error)
-  
-  useEffect(() => { 
-    if ((entity?.name !== entityName || !entity?.properties?.length) && !loading && !error) {
-      getEntity(entityName)
-    } 
-  }, [entityName, loading, error, entity])
-
-  return entity
-    ? { 
-      ...entity,
-      fields: mapFieldComponents(entity),
-      label: entity.layout.label || entityName
-    }
-    : { 
-      name: entityName,
-      layout: { label: entityName },
-      label: entityName, 
-      properties: [],
-      fields: []
-    }
 }
